refactor: move `get_distribution_roots_length` to reader
